refactor: Move expectApprox to core-utils
