+++ packages/core-utils/src/common/index.ts
@@ -1,2 +1,3 @@
 export * from './hex-strings'
 export * from './misc'
+export * from './test-utils'
